update
I definitely do not want to do that over again. wow. that was
intensive. but at least I got rid of the hardest part. the rest that’s
left is to finish the easy part for the motorcycle video page, and then
to implement the navigation once the index.php is slightly reworked.
place in the footer function, then stylize the navigational area.
hopefully that will be by 2am once i’m finished. thursday will be
creating all the assets needed, and friday will be working on the piece
of the pie, rebrand party.